% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_template.R
\name{setup_folders}
\alias{setup_folders}
\title{Command to set up an optimized folder structure for your project}
\usage{
setup_folders()
}
\description{
Folders that will be generated matching the Post's standards
}
\examples{
setup_folders()
}
